requirements.txt
Pillow
pyfiglet
numpy